Server/requirements.txt
asgiref==3.6.0
attrs==23.1.0
autobahn==23.1.2
Automat==22.10.0
beautifulsoup4==4.11.2
certifi==2023.5.7
cffi==1.15.1
channels==4.0.0
charset-normalizer==3.1.0
click==8.1.3
constantly==15.1.0
contourpy==1.0.7
coreapi==2.3.3
coreschema==0.0.4
cryptography==40.0.2
cycler==0.11.0
daphne==4.0.0
Django==3.2.18
django-cors-headers==3.14.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.5
fonttools==4.39.4
gunicorn==20.1.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
hyperframe==6.0.1
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
inflection==0.5.1
itypes==1.2.0
jieba==0.42.1
Jinja2==3.1.2
kiwisolver==1.4.4
MarkupSafe==2.1.2
matplotlib==3.7.1
numpy==1.24.3
packaging==23.1
Pillow==9.4.0
priority==1.3.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.7.0
pyOpenSSL==23.1.1
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
requests==2.28.2
ruamel.yaml==0.17.26
ruamel.yaml.clib==0.2.7
service-identity==21.1.0
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.4
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.5.0
uritemplate==4.1.1
urllib3==1.26.15
uvicorn==0.21.1
wordcloud==1.9.1.1
zope.interface==6.0
